+++ src/esm/utils/datastructures/ColorGradient.js
@@ -25,12 +25,12 @@ export class ColorGradient {
      * order regarding the `ratio` information. This will not be validated.
      *
      * @param {Array<ColorGradientItem>} values - The values to use.
-     * @param {number?} angleInRadians - (optional) An optional angle for the gradient; if no specified `PI/2.0` will be used (vertical from left to right).
+     * @param {number?} angle - (optional) An optional angle for the gradient; if no specified `PI/2.0` will be used (vertical from left to right).
      */
-    constructor(values, angleInRadians) {
+    constructor(values, angle) {
         this.values = values;
         // Default: left to right
-        this.angle = typeof angleInRadians === "undefined" ? Math.PI / 2.0 : angleInRadians;
+        this.angle = typeof angle === "undefined" ? Math.PI / 2.0 : angle;
     }
     /**
      * Get a color gradient CSS value string from these gradient settings.
@@ -42,7 +42,10 @@ export class ColorGradient {
     toColorGradientString() {
         // Example:
         //    linear-gradient(90deg,rgba(42, 123, 155, 1) 0%, rgba(87, 199, 133, 1) 38%, rgba(161, 210, 108, 1) 68%, rgba(237, 221, 83, 1) 100%)
-        const buffer = [`linear-gradient( ${RAD_TO_DEG * this.angle}deg, `];
+        // const buffer: Array<string | undefined> = [`linear-gradient( ${RAD_TO_DEG * this.angle}deg, `];
+        const buffer = [
+            `linear-gradient( ${typeof this.angle === "number" ? `${this.angle * RAD_TO_DEG}deg` : this.angle}, `
+        ];
         for (var i = 0; i < this.values.length; i++) {
             if (i > 0) {
                 buffer.push(",");
@@ -50,7 +53,7 @@ export class ColorGradient {
             const colorValue = this.values[i] ? this.values[i].color.cssRGBA() : null;
             buffer.push(colorValue);
             const percentage = this.values[i].ratio;
-            buffer.push(`${percentage * 100}%`);
+            buffer.push(`${typeof percentage === "number" ? `${percentage * 100}%` : percentage}`);
         }
         buffer.push(")");
         return buffer.join(" ");
@@ -96,6 +99,37 @@ export class ColorGradient {
             }
         }
         return [leftSliderIndex, closestSliderValue];
+    }
+    /**
+     * Try to convert the given anonymous item (usually and object) to a ColorGradient. This method should be used
+     * to restore items from JSON parse results to convert them into proper ColorGradient instances.
+     *
+     * @param {any} item
+     * @returns {ColorGradient}
+     */
+    static fromObject(item) {
+        if (typeof item !== "object") {
+            throw new Error(`Cannot instantiate ColorGradient from item: must be an object. (${typeof item})`);
+        }
+        if (!item.hasOwnProperty("values")) {
+            throw new Error("Cannot instantiate ColorGradient from object: missing property 'values'.");
+        }
+        const valuesObj = item.values;
+        if (!Array.isArray(valuesObj)) {
+            throw new Error("Cannot instantiate ColorGradient from object: property 'values' is not an array.");
+        }
+        const values = valuesObj.map((element, index) => {
+            if (typeof element !== "object") {
+                throw new Error(`Cannot instantiate ColorGradient from object: item at index ${index} is not an object.`);
+            }
+            if (!element.hasOwnProperty("color") || !element.hasOwnProperty("ratio")) {
+                throw new Error(`Cannot instantiate ColorGradient from object: item at index ${index} is missing 'color' or 'ratio' property.`);
+            }
+            const tmpElement = element;
+            return { color: Color.makeRGB(tmpElement.color.r, tmpElement.color.g, tmpElement.color.b), ratio: tmpElement.ratio };
+        });
+        // More checks possible here ...
+        return new ColorGradient(values, item.angle);
     }
     /**
      * Clone this linear color gradient. Returns a deep clone.