+++ src/esm/utils/dom/components/ColorGradientSelector.js
@@ -86,19 +86,25 @@ export class ColorGradientSelector {
         this.container.append(this._render());
     } // END constructor
     // TODO: implement this?
-    // public setGradients(gradients: Array<ColorGradient>, selectedIndex: number): void {
-    //   // First empty the container
-    //   this.__removeAllChildNodes(this.positioningContainerRef.current);
-    //   // Clone the array
-    //   this.colorGradients = gradients.map((gradient: ColorGradient) => gradient);
-    //   // Re-render button array
-    //   // const newChildNodes: Array<NoReact.Ref<HTMLButtonElement>> = this._renderAllOptionButtons();
-    //   this._renderAllOptionButtons();
-    //   // Re-fill container with new nodes
-    //   this.colorGradientOptionRefs.forEach((nodeRef: NoReact.Ref<HTMLButtonElement>) => {
-    //     this.positioningContainerRef.current.appendChild(nodeRef.current);
-    //   });
-    // }
+    setGradients(gradients, selectedIndex) {
+        // First empty the container
+        this.__removeAllChildNodes(this.positioningContainerRef.current);
+        // Clone the array
+        // this.colorGradients = gradients.map((gradient: ColorGradient) => gradient);
+        this.colorGradients = [];
+        this.colorGradientOptionRefs = [];
+        this.selectedGradientIndex = selectedIndex;
+        // Re-render button array
+        // const newChildNodes: Array<NoReact.Ref<HTMLButtonElement>> = this._renderAllOptionButtons();
+        // this._renderAllOptionButtons();
+        // Re-fill container with new nodes
+        // this.colorGradientOptionRefs.forEach((nodeRef) => {
+        //   this.positioningContainerRef.current.appendChild(nodeRef.current);
+        // });
+        gradients.map((grad) => {
+            this.addGradient(grad);
+        });
+    }
     addGradient(gradient) {
         console.log("addGradient");
         // Render a new button
@@ -323,7 +329,32 @@ export class ColorGradientSelector {
                     l: 0,
                     zIndex: 999
                 } }, _self._renderAllOptionButtons())));
-    } // END functionrender()
+    } // END function render()
+    storeInLocalStorage() {
+        // Convert gradients to array of strings
+        // const arr: Array<string> = this.colorGradients.map((grad: ColorGradient) => {
+        //   // return grad.toColorGradientString();
+        //   return JSON.stringify(grad);
+        // });
+        // localStorage.setItem("ColorGradientSelector", JSON.stringify(arr));
+        localStorage.setItem("ColorGradientSelector", JSON.stringify(this.colorGradients));
+    }
+    restoreFromLocalStorage() {
+        // Storage value to array of ColorGradients
+        const value = localStorage.getItem("ColorGradientSelector");
+        if (!value) {
+            return;
+        }
+        const jsonArray = JSON.parse(value);
+        if (!Array.isArray(jsonArray)) {
+            return;
+        }
+        const gradArray = jsonArray.map((gradientObject) => {
+            console.log("Converting object to ColorGradient", gradientObject);
+            return ColorGradient.fromObject(gradientObject);
+        });
+        this.setGradients(gradArray, 0);
+    }
 }
 ColorGradientSelector.DEFAULT_COLOR_GRADIENTS = [
     ColorGradient.createDefault(),